Added a basic lexer.
Split the input into a sequence of tokens, where each token is either
a nonempty sequence of alpha-numeric char or one of the parens (,).
Parsing the sequence of tokens is easier that parsing a raw string:
we don't need to worry about stripping the white space, and we don't
have overlapping patterns (like "if" and "iszero").  Furthermore,
error messages are also improved.  Note that parsing the program from
the sequence of tokens requires a small hack: we need to reset the
initial position (which is line 1, column 1) to the position of the
first token in the sequence, otherwise if there is an error in the
first token, its location won't be reported correctly.