Added basic primitives: exp, sin, and cos (for expository purposes).
It seems that the only sensible thing to do is to define D f u, where
f is a primitive, as \x . (f' x) * u, where f' is the derivative of f
and x is a fresh variable (not occurring free in u).  On the other
hand, the partial derivative of a primitive is a function that
constantly returns 0.  These seem to play nicely together.