Added a simple parser and an interactive interpreter.
To simplify the parser, I have chosen to make the language
fully-parenthesized, like Lisp.  The main difficulty is parsing of
Doubles, and for that I use Token and Language modules.  The lexer is
pretty basic and doesn't keep track of the position of a token; I am
going to change that soon.