Added a reduction rule for numbers in the function position.
A number is punned with the function that multiplies its argument with
that number.  This seems a sensible thing to do, and furthermore it is
necessary to get `partial' right: (dy/dx)u is defined as 0 when y is a
variable different from x, however if y is bound to a function in some
environment, then (dy/dx)u must be able to act like a function.

With the current definition Church encoded pairs seem to work:

> (((derive (lambda x (lambda f ((f x) (* x x)))) 1) 2) (lambda x (lambda y y)))
4.0
> (((derive (lambda x (lambda f ((f x) (* x x)))) 1) 2) (lambda x (lambda y x)))
1.0
> (((derive (lambda x (lambda f ((f (sin x)) (cos x)))) 1) 3.141592653589793) (lambda x (lambda y x)))
-1.0
> (((derive (lambda x (lambda f ((f (sin x)) (cos x)))) 1) 3.141592653589793) (lambda x (lambda y y)))
-1.2246467991473532e-16